+++ fsspec/implementations/tests/test_jupyter.py
@@ -1,5 +1,4 @@
 import os
-import re
 import shlex
 import subprocess
 import time
@@ -13,37 +12,27 @@ requests = pytest.importorskip("requests")
 
 @pytest.fixture()
 def jupyter(tmpdir):
-    import requests
 
     tmpdir = str(tmpdir)
+    os.environ["JUPYTER_TOKEN"] = "blah"
     try:
-        P = subprocess.Popen(
-            shlex.split(
-                f"jupyter notebook --notebook-dir={tmpdir}" f" --no-browser --port=5566"
-            ),
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            stdin=subprocess.DEVNULL,
-        )
+        cmd = f"jupyter notebook --notebook-dir={tmpdir} --no-browser --port=5566"
+        P = subprocess.Popen(shlex.split(cmd))
     except FileNotFoundError:
         pytest.skip("notebook not installed correctly")
     try:
-        timeout = 5
+        timeout = 15
         while True:
             try:
-                r = requests.get("http://127.0.0.1:5566/")
+                r = requests.get("http://localhost:5566/?token=blah")
                 r.raise_for_status()
                 break
             except (requests.exceptions.BaseHTTPError, IOError):
                 time.sleep(0.1)
                 timeout -= 0.1
-                pytest.skip("Timed out for jupyter")
-            txt = P.stdout.read(600).decode()
-            try:
-                url = re.findall("(http[s]*://[^\\n]+)", txt)[0]
-            except IndexError:
-                pytest.skip("No notebook URL: " + txt)  # debug on fail
-            yield url, tmpdir
+                if timeout < 0:
+                    pytest.xfail("Timed out for jupyter")
+        yield "http://localhost:5566/?token=blah", tmpdir
     finally:
         P.terminate()