+++ fsspec/implementations/tests/test_libarchive.py
@@ -1,20 +1,20 @@
-import zipfile
 from contextlib import contextmanager
 
 import os
 import pickle
 import pytest
-import sys
 import tempfile
 import fsspec
 
+libarchive = pytest.importorskip("libarchive")
+
 
 @contextmanager
-def tempzip(data={}):
-    f = tempfile.mkstemp(suffix="zip")[1]
-    with zipfile.ZipFile(f, mode="w") as z:
+def temparchive(data={}):
+    f = tempfile.mkstemp(suffix="7z")[1]
+    with libarchive.file_writer(f, "7zip") as archive:
         for k, v in data.items():
-            z.writestr(k, v)
+            archive.add_file_from_memory(entry_path=k, entry_size=len(v), entry_data=v)
     try:
         yield f
     finally:
@@ -28,8 +28,8 @@ data = {"a": b"", "b": b"hello", "deeply/nested/path": b"stuff"}
 
 
 def test_empty():
-    with tempzip() as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive() as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
         assert fs.find("") == []
         assert fs.find("", withdirs=True) == []
         with pytest.raises(FileNotFoundError):
@@ -37,32 +37,29 @@ def test_empty():
         assert fs.ls("") == []
 
 
-def test_glob():
-    with tempzip(data) as z:
-        fs = fsspec.filesystem("zip", fo=z)
-        assert fs.glob("*/*/*th") == ["deeply/nested/path"]
-
-
-@pytest.mark.xfail(sys.version_info < (3, 6), reason="zip-info odd on py35")
 def test_mapping():
-    with tempzip(data) as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
         m = fs.get_mapper("")
+
+        fs._get_dirs()
+        print(fs.dir_cache)
+
         assert list(m) == ["a", "b", "deeply/nested/path"]
         assert m["b"] == data["b"]
 
 
-@pytest.mark.xfail(sys.version_info < (3, 6), reason="zip not supported on py35")
 def test_pickle():
-    with tempzip(data) as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
         fs2 = pickle.loads(pickle.dumps(fs))
+        assert fs2 is fs
         assert fs2.cat("b") == b"hello"
 
 
 def test_all_dirnames():
-    with tempzip() as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive() as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
 
         # fx are files, dx are a directories
         assert fs._all_dirnames([]) == set()
@@ -75,8 +72,8 @@ def test_all_dirnames():
 
 
 def test_ls():
-    with tempzip(data) as z:
-        lhs = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        lhs = fsspec.filesystem("libarchive", fo=archive_file)
 
         assert lhs.ls("") == ["a", "b", "deeply/"]
         assert lhs.ls("/") == lhs.ls("")
@@ -89,8 +86,8 @@ def test_ls():
 
 
 def test_find():
-    with tempzip(data) as z:
-        lhs = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        lhs = fsspec.filesystem("libarchive", fo=archive_file)
 
         assert lhs.find("") == ["a", "b", "deeply/nested/path"]
         assert lhs.find("", withdirs=True) == [
@@ -106,26 +103,30 @@ def test_find():
 
 
 def test_walk():
-    with tempzip(data) as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
         expected = [
             # (dirname, list of subdirs, list of files)
             ("", ["deeply"], ["a", "b"]),
             ("deeply", ["nested"], []),
             ("deeply/nested", [], ["path"]),
         ]
-        assert list(fs.walk("")) == expected
+        for lhs, rhs in zip(fs.walk(""), expected):
+            assert lhs[0] == rhs[0]
+            assert sorted(lhs[1]) == sorted(rhs[1])
+            assert sorted(lhs[2]) == sorted(rhs[2])
 
 
 def test_info():
-    with tempzip(data) as z:
-        fs_cache = fsspec.filesystem("zip", fo=z)
+    with temparchive(data) as archive_file:
+        fs_cache = fsspec.filesystem("libarchive", fo=archive_file)
 
         with pytest.raises(FileNotFoundError):
             fs_cache.info("i-do-not-exist")
 
         # Iterate over all directories
-        # The ZipFile does not include additional information about the directories,
+        # The 7zip archive does not include additional information about the
+        # directories
         for d in fs_cache._all_dirnames(data.keys()):
             lhs = fs_cache.info(d)
             expected = {"name": f"{d}/", "size": 0, "type": "directory"}
@@ -138,10 +139,12 @@ def test_info():
             assert lhs["size"] == len(v)
             assert lhs["type"] == "file"
 
-            # There are many flags specific to Zip Files.
-            # These are two we can use to check we are getting some of them
-            assert "CRC" in lhs
-            assert "compress_size" in lhs
+            # These are the specific flags retrieved from the archived files
+            assert "created" in lhs
+            assert "mode" in lhs
+            assert "uid" in lhs
+            assert "gid" in lhs
+            assert "mtime" in lhs
 
 
 @pytest.mark.parametrize("scale", [128, 512, 4096])
@@ -152,8 +155,8 @@ def test_isdir_isfile(scale):
         return "/".join(x.translate(table))
 
     scaled_data = {f"{make_nested_dir(i)}/{i}": b"" for i in range(1, scale + 1)}
-    with tempzip(scaled_data) as z:
-        fs = fsspec.filesystem("zip", fo=z)
+    with temparchive(scaled_data) as archive_file:
+        fs = fsspec.filesystem("libarchive", fo=archive_file)
 
         lhs_dirs, lhs_files = fs._all_dirnames(scaled_data.keys()), scaled_data.keys()